main_dir/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-filter==23.2
psycopg2-binary==2.9.6
geopy==2.3.0
celery==5.2.7
django-celery-beat==2.5.0
redis==4.5.5
